PCHR-3142: Arrange backstop tests for contact actions
