+++ uk.co.compucorp.civicrm.hrcore/backstop_data/casper_scripts/contact-summary/actions/show-user-account-delete-confirmation-dialog.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var page = require('../../page-objects/contact-actions-menu');
+var page = require('../../../page-objects/contact-summary');
 
 module.exports = function (casper) {
   page.init(casper).deleteUserAccount();
