--- uk.co.compucorp.civicrm.hrcore/backstop_data/scenarios/contact-actions-menu.json
@@ -1,16 +1,0 @@
-{
-  "scenarios": [
-    {
-      "label": "Contact Summary - Action Menu",
-      "url": "index.php?q=civicrm/contact/view&reset=1&cid={{contactId:staff}}",
-      "onReadyScript": "contact/open-contact-actions-menu",
-      "credential": "admin"
-    },
-    {
-      "label": "Contact Action Menu - User deletetion confirmation dialog",
-      "url": "index.php?q=civicrm/contact/view&reset=1&cid={{contactId:staff}}",
-      "onReadyScript": "contact/delete-user-account-confirmation-dialog",
-      "credential": "admin"
-    }
-  ]
-}
